// client/src/Frontend/Components/CoreUI.tsx
export const Spacer = styled.div`
  ${({ width, height }: { width?: number; height?: number }) => css`